+++ app/policies/vye_policy.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 VyePolicy = Struct.new(:user, :user_info) do
   def access?
     return true if user.present?
