+++ modules/vye/lib/dgib/service.rb
@@ -2,6 +2,10 @@
 
 require 'dgi/configuration'
 require 'common/client/base'
+require_relative 'claimant_lookup_response'
+require_relative 'claimant_status_response'
+require_relative 'verification_record_response'
+require_relative 'verify_claimant_response'
 
 module Vye
   module DGIB
@@ -23,6 +27,90 @@ module Vye
             key
           end
         end
+      end
+
+      def claimant_lookup(ssn)
+        params = ActionController::Parameters.new({ ssn: })
+        with_monitoring do
+          headers = request_headers
+          options = { timeout: 60 }
+          response = perform(:post, claimant_lookup_end_point, camelize_keys_for_java_service(params).to_json, headers,
+                             options)
+          Vye::DGIB::ClaimantLookupResponse.new(response.status, response)
+        end
+      end
+
+      def get_claimant_status(claimant_id)
+        with_monitoring do
+          headers = request_headers
+          options = { timeout: 60 }
+          raw_response = perform(:get, claimant_status_end_point(claimant_id), {}, headers, options)
+          Vye::DGIB::ClaimantStatusRecordResponse.new(raw_response.status, raw_response)
+        end
+      end
+
+      def verify_claimant(
+        claimant_id,
+        verified_period_begin_date,
+        verified_period_end_date,
+        verified_through_date,
+        verification_method,
+        response_type
+      )
+
+        params = ActionController::Parameters.new({
+                                                    claimant_id:,
+                                                    verified_period_begin_date:,
+                                                    verified_period_end_date:,
+                                                    verified_through_date:,
+                                                    verification_method:,
+                                                    app_communication: { response_type: }
+                                                  })
+
+        with_monitoring do
+          headers = request_headers
+          options = { timeout: 60 }
+          response = perform(:post, verify_claimant_end_point, camelize_keys_for_java_service(params).to_json, headers,
+                             options)
+          Vye::DGIB::VerifyClaimant::Response.new(response.status, response)
+        end
+      end
+
+      def get_verification_record(claimant_id)
+        with_monitoring do
+          headers = request_headers
+          options = { timeout: 60 }
+          raw_response = perform(:get, verification_record_end_point(claimant_id), {}, headers, options)
+          Vye::DGIB::VerificationRecordResponse.new(raw_response.status, raw_response)
+        end
+      end
+
+      private
+
+      def claimant_lookup_end_point
+        'dgi/vye/claimantLookup'
+      end
+
+      def claimant_status_end_point(claimant_id)
+        "verifications/vye/#{claimant_id}/status"
+      end
+
+      def verify_claimant_end_point
+        'verifications/vye/verify'
+      end
+
+      def verification_record_end_point(claimant_id)
+        "verifications/vye/#{claimant_id}/verification-record"
+      end
+
+      def json
+        nil
+      end
+
+      def request_headers
+        {
+          Authorization: "Bearer #{DGIB::AuthenticationTokenService.call}"
+        }
       end
     end
   end
