--- modules/vye/lib/dgib/verify_claimant/configuration.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require 'dgib/configuration'
-
-module Vye
-  module DGIB
-    module VerifyClaimant
-      class Configuration < Vye::DGIB::Configuration
-        def service_name
-          'DGIB/VerifyClaimant'
-        end
-      end
-    end
-  end
-end
